website_crew/requirements.txt
crewai>=0.80.0
crewai-tools>=0.14.0
python-dotenv>=1.0.0
pyyaml>=6.0
